2015-08-27 Fred Gleason <fredg@paravelsystems.com> 	* Added an '--asihpi-input-gain' switch to the ASIHPI source in 	'src/glasscoder/asihpidevice.cpp' and 'src/glasscoder/asihpidevice.h'. 	* Added an '--asihpi-channel-mode' switch to the ASIHPI source in 	'src/glasscoder/asihpidevice.cpp' and 'src/glasscoder/asihpidevice.h'.

+++ src/glassgui/hpiinputlistview.cpp
@@ -98,8 +98,8 @@ unsigned HpiInputListView::selectedAdapterIndex() const
   if(!currentIndex().isValid()) {
     return 0;
   }
-  return hpi_model->data(currentIndex(),HpiInputListViewModel::AdapterRole).
-    toUInt();
+  return hpi_model->data(currentIndex(),
+			 HpiInputListViewModel::AdapterRole).toUInt();
 }
 
 
@@ -108,8 +108,8 @@ unsigned HpiInputListView::selectedInputIndex() const
   if(!currentIndex().isValid()) {
     return 0;
   }
-  return hpi_model->data(currentIndex(),HpiInputListViewModel::InputRole).
-    toUInt();
+  return hpi_model->data(currentIndex(),
+			 HpiInputListViewModel::InputRole).toUInt();
 }
